refactor: decompose membership-list into focused sub-components (#58)
Extract active-member-list (search + archive/role mutations) and
add-member-form (Superforms + candidates query) from the monolith
membership-list component, aligning the add-member form with project
Superforms/Formsnap/Zod conventions.
