+++ src/routes/app/admin/components/membership-list/active-member-list/index.ts
@@ -1,0 +1,1 @@
+export { default as ActiveMemberList } from './active-member-list.svelte';
