+++ src/routes/app/admin/components/membership-list/add-member-form/index.ts
@@ -1,0 +1,1 @@
+export { default as AddMemberForm } from './add-member-form.svelte';
